backend/requirements.txt
# requirements.txt for installing necessary dependencies
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.6.0
psycopg[binary]
Pillow==10.4.0
moviepy==2.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
trimesh
numpy==2.2.1